cmd/commands package * moved to cmd
