+++ cmd/exec_test.go
