+++ cmd/testdata/mta/node-js/gulpfile.js
