+++ cmd/testdata/mta/node-js/package.json
