+++ cmd/testdata/mtahtml5/testapp/webapp/Component.js
