+++ cmd/testdata/mtahtml5/testapp/webapp/controller/View1.controller.js
