+++ cmd/testdata/mtahtml5/testapp/webapp/view/View1.view.xml
